requirements.txt
flask==3.1.0
flask-cors==5.0.1
rembg[cpu]==2.0.62
Pillow==11.1.0
